fix: fix the terminator unit test

+++ mava/utils/lp_utils.py
@@ -25,6 +25,7 @@ from absl import flags, logging
 from acme.utils import counting
 from launchpad.nodes.python.local_multi_processing import PythonProcess
 
+from mava.core_jax import SystemParameterServer
 from mava.utils.training_utils import non_blocking_sleep
 
 FLAGS = flags.FLAGS
@@ -87,15 +88,22 @@ def partial_kwargs(function: Callable[..., Any], **kwargs: Any) -> Callable[...,
     return functools.partial(function, **kwargs)
 
 
-def termination_fn(parent_pid: int) -> None:
+def termination_fn(
+    parameter_server: SystemParameterServer,
+) -> None:
     """Terminate the process
 
     Args:
-        parent_pid: the pid of the main thread process
+        parameter_server: SystemParameterServer in order to get main pid
     """
-    parent = psutil.Process(parent_pid)
-    for child in parent.children(recursive=True):
-        child.kill()
+    if parameter_server.store.manager_pid:
+        # parent_pid: the pid of the main thread process
+        parent_pid = parameter_server.store.manager_pid
+        parent = psutil.Process(parent_pid)
+        for child in parent.children(recursive=True):
+            child.kill()
+    else:
+        lp.stop()
 
 
 class StepsLimiter:
